Move the error-dialog decorator back to microdrop_utils (review)
Per review on #459: the decorator returns to
microdrop_utils/decorators.py as before. It keeps the shared dialog
helpers + DIALOG_ERROR_TEXT_COLOR from the dedupe pass; its unit tests
land in a new microdrop_utils/tests/ (their subject lives here, not in
the tree plugin or the dialogs package).
